pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spe-sc"
version = "1.0.0"
description = "SPE-SC: A defense mechanism for Large Language Models against adversarial attacks"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "SPE-SC Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "langchain>=0.2.14",
    "langchain-community>=0.2.14",
    "langchain-openai>=0.2.14",
    "langchain-huggingface>=0.2.14",
    "langchain-xai",
    "langchain_azure_ai",
    "pydantic>=2.0.1",
    "PyYAML>=6.0",
    "loguru>=0.7.3",
    "rouge-score>=0.1.2",
    "torch>=2.0.0",
    "nltk>=3.8.1",
    "pandas>=2.1.1",
    "openpyxl",
    "sentence-transformers",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black",
    "flake8",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."
"src" = "src"
